python-services/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg[binary]==3.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.8.2
